VRIOT/authServer/requirements.txt
appdirs==1.4.0
Django==1.10.5
django-filter==1.0.1
djangorestframework==3.5.3
Markdown==2.6.7
mongoengine==0.11.0
packaging==16.8
pymongo==3.4.0
pyparsing==2.1.10
six==1.10.0
